+++ core_gearworks_lib/source/lib/presets/2dim/rectangle/p2d_rectangle.h
@@ -16,11 +16,13 @@ public:
 	/// A blend colour and a texture may be given.
 	/// </summary>
 	/// <param name="renderer">The renderer that the rectangle will be 'bound' to.</param>
+	/// <param name="texture_path">The texture to use for the rectangle.</param>
 	/// <param name="position">The starting position coordinate of the rectangle primitive, with 0 as its center.</param>
 	/// <param name="scale">The scale values of the rectangle primitive.</param>
-	/// <param name="texture_path">The texture to use for the rectangle.</param>
+	/// <param name="rotation">The starting rotation of the rectangle.</param>
 	/// <param name="blend_colour">The colour to render the rectangle. Default is white.</param>
-	void create(gearworks::renderer *renderer, vec3 position, vec2 scale, string texture_path, vec3 blend_colour = vec3(255, 255, 255), float blend_opacity = 255);
+	/// <param name="blend_opacity">The opacity of the shape.</param>
+	void create(gearworks::renderer *renderer, string texture_path, vec3 position = vec3(0), vec2 scale = vec2(1), vec3 rotation = vec3(0), vec3 blend_colour = vec3(255, 255, 255), float blend_opacity = 255);
 };
 
 #endif // header guard