+++ demo/source/win32_process_main.cpp
@@ -4,21 +4,19 @@ gearworks::rectangle rect;
 gearworks::triangle tri;
 
 void gearworks::engine::initialize() {
-	rect.create(renderer, vec3(0, 0, 500), vec2(400, 250), "resources/textures/crate.png");
-	rect.rotate(0, 90, 0);
+	rect.create(renderer, "resources/textures/crate.png", vec3(0, 0, 500), vec2(100), vec3(0, 0, 0));
 
 	vec3 vertices[3] = {
 		vec3(-100, -200, 0),
-		vec3(0, 200, 0),
+		vec3(100, 100, 0),
 		vec3(100, -200, 0)
 	};
 
-	tri.create(renderer, vertices, "resources/textures/wall.jpeg");
-	tri.translate(0, 0, 500);
+	tri.create(renderer, GEAR_RIGHTANGLE_TRI, "resources/textures/wall.jpeg", vec3(0, 0, 500), vec2(100));
 }
 void gearworks::engine::update() {
 	rect.rotate(0, 1, 0);
-	tri.rotate(0, -2, 0);
+	tri.rotate(0, -1, 0);
 
 	if (input->get_key(GLFW_KEY_W)) { renderer->view_matrix = glm::rotate(renderer->view_matrix, glm::radians(1.0f), vec3(1, 0, 0)); }
 	if (input->get_key(GLFW_KEY_A)) { renderer->view_matrix = glm::rotate(renderer->view_matrix, glm::radians(-1.0f), vec3(0, 1, 0)); }
@@ -29,8 +27,8 @@ void gearworks::engine::render() {
 	// Clears the screen to a dark grey colour.
 	gearworks::clear_screen(gearworks::col_verydarkgray);
 
-	tri.render();
 	rect.render();
+	tri.render();
 }
 void gearworks::engine::clean() 
 {}